Some more customer edits, added employee UI functionality and resizing stuff. Changed calcTicketCost to return double because error will robinson

+++ AirEase-AMS/App/Ticket/Ticket.cs
@@ -162,7 +162,7 @@ public class Ticket : ITicket
 
 
     
-    public decimal CalculateTicketCost()
+    public double CalculateTicketCost()
     {
         if (CalculateStraightLineMileage() == 0) return 0;
         return Convert.ToDouble(flights[0].CalculateFlightCost()) + (8 * num_flights);
